corpus: move all npbench+polybench computation to fp64/complex128
Bump every npbench and polybench kernel from float32/complex64 to
float64/complex128 (dc_float, the numpy initialize() datatype, and the two
hardcoded np.float32 buffers in cavity_flow). Both the dace kernel and its numpy
reference move together, so the value-preserving corpus comparisons run at fp64
where float32 reassociation noise was previously pushing chaotic kernels
(cavity_flow, the mandelbrot escape counts, stockham_fft) onto the tolerance
edge. tsvc is already fp64; the inert #datatypes menu comment is left as-is.

+++ tests/corpus/npbench/combinational_logic/crc16.py
@@ -4,8 +4,8 @@ import numpy as np
 import dace
 import dace as dc
 
-dc_float = dc.float32
-dc_complex_float = dc.complex64
+dc_float = dc.float64
+dc_complex_float = dc.complex128
 
 SIZES = {'N': 1600}
 INPUT_ARGS = ('N', )
