+++ tests/corpus/npbench/dense_linear_algebra/permute_3d.py
@@ -3,8 +3,8 @@
 import numpy as np
 import dace as dc
 
-dc_float = dc.float32
-dc_complex_float = dc.complex64
+dc_float = dc.float64
+dc_complex_float = dc.complex128
 
 SIZES = {'N': 128}
 INPUT_ARGS = ('N', )
@@ -15,7 +15,7 @@ OUTPUT_ARGS = ('B', )
 N = dc.symbol('N', dtype=dc.int64)
 
 
-def initialize(N, datatype=np.float32):
+def initialize(N, datatype=np.float64):
     A = np.fromfunction(lambda i, j, k: (i * N * N + j * N + k * N) / N, (N, N, N), dtype=datatype)
     B = np.zeros((N, N, N), dtype=datatype)
     return (A, B)
