+++ tests/corpus/npbench/ml/lenet.py
@@ -3,8 +3,8 @@
 import numpy as np
 import dace as dc
 
-dc_float = dc.float32
-dc_complex_float = dc.complex64
+dc_float = dc.float64
+dc_complex_float = dc.complex128
 
 # C_before_fc1 = 16 * (H_pool2 * W_pool2). The dace kernel uses it as a free symbol
 # in ``np.reshape(x, (N, C_before_fc1))`` that cannot be inferred from a transient,
@@ -24,7 +24,7 @@ N, H, W, C_before_fc1, S0, S1, S2, S3, S4, S5 = (dc.symbol(s, dtype=dc.int64)
                                                            'S5'))
 
 
-def initialize(N, H, W, datatype=np.float32):
+def initialize(N, H, W, datatype=np.float64):
     from numpy.random import default_rng
     rng = default_rng(42)
     H_conv1 = H - 4
